Move model factory to run models
Model factory is used by both cli and gui.
