+++ src/ert/cli/main.py
@@ -9,7 +9,6 @@ import sys
 from typing import Optional, TextIO
 
 from _ert.threading import ErtThread
-from ert.cli.model_factory import create_model
 from ert.cli.monitor import Monitor
 from ert.cli.workflow import execute_workflow
 from ert.config import ErtConfig, QueueSystem
@@ -25,6 +24,7 @@ from ert.mode_definitions import (
 )
 from ert.namespace import Namespace
 from ert.run_models.base_run_model import StatusEvents
+from ert.run_models.model_factory import create_model
 from ert.shared.plugins import ErtPluginManager
 from ert.storage import open_storage
 from ert.storage.local_storage import local_storage_set_ert_config
