+++ src/ert/gui/simulation/experiment_panel.py
@@ -20,7 +20,6 @@ from qtpy.QtWidgets import (
     QWidget,
 )
 
-from ert.cli.model_factory import create_model
 from ert.enkf_main import EnKFMain
 from ert.gui.ertnotifier import ErtNotifier
 from ert.libres_facade import LibresFacade
@@ -29,6 +28,7 @@ from ert.mode_definitions import (
     ES_MDA_MODE,
     ITERATIVE_ENSEMBLE_SMOOTHER_MODE,
 )
+from ert.run_models.model_factory import create_model
 
 from .ensemble_experiment_panel import EnsembleExperimentPanel
 from .ensemble_smoother_panel import EnsembleSmootherPanel
